+++ NEWS.md
@@ -1,6 +1,7 @@
 27.3 (Not released yet)
 =================
 
+* Introduce a notification system (`--notify-events` testsuite argument).
 * Add a `--list-json=FILE` CLI argument to dump the list of tests in a
   testsuite to a file in JSON format.